+++ src/types/flags.ts
@@ -1,0 +1,3 @@
+import type { CountryCode as _CountryCode } from 'libphonenumber-js';
+
+export type CountryCode = _CountryCode | 'AST';